+++ src/app/(normal)/detail/[platform]/[id]/loading.tsx
