+++ src/app/(normal)/profile/components/ContentCollected.tsx
