+++ src/app/(normal)/profile/components/PlatformIcon.tsx
