+++ src/app/(normal)/profile/components/PlatformTabs.tsx
