+++ src/app/layout.tsx
@@ -10,7 +10,6 @@ import { Polyfills } from '@/components/Polyfills.js';
 import { SearchBar } from '@/components/SearchBar.js';
 import { SearchFilter } from '@/components/SearchFilter.js';
 import { SideBar } from '@/components/SideBar/index.js';
-import { SocialPlatformTabs } from '@/components/SocialPlatformTabs.js';
 import { createSiteMetadata } from '@/helpers/createSiteMetadata.js';
 import { Modals } from '@/modals/index.js';
 
@@ -29,17 +28,9 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
                 <Providers>
                     <div className="m-auto flex min-h-screen sm:w-full lg:w-[1265px] ">
                         <SideBar />
-
-                        <main className="max-w-[888px] flex-1 border-r border-line pl-72">
-                            <div className="sticky top-0 z-[998] bg-white dark:bg-black">
-                                <SearchBar source="header" />
-
-                                <SocialPlatformTabs />
-                            </div>
-
+                        <main>
                             {children}
                         </main>
-
                         <aside className=" sticky top-0 z-[998] h-full w-96 px-4 lg:block">
                             <SearchBar source="secondary" />
 
@@ -52,8 +43,8 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
                     <PageInspectorRender />
                 </Providers>
                 <GA />
-            </body>
+            </body >
             <ScrollRestorer />
-        </html>
+        </html >
     );
 }
